+++ eva/src/index.js
@@ -167,6 +167,8 @@ Vue.component('dash-scatterPlot', () => import('./components/dashBoardInside/das
 Vue.component('dash-frequencyGraph', () => import('./components/dashBoardInside/dashFrequencyGraph.vue'));
 // Компонент редактора кода
 Vue.component('code-editor', () => import('./components/codeEditor.vue'));
+// Спидометр
+Vue.component('dash-gauge', () => import('./components/dashBoardInside/dashGauge/dashGauge'));
 // убирает фокус с элемена сразу либо через переданное ms v-blur="140"
 Vue.directive('blur', {
   inserted(el, binding) {